`notebook [name] <image>` makes a jupyterhub notebook.
`start` fires it up onto port 8888, more `start` s give you bash sessions.

Still needs a bit of `docker-machine ssh` in `start` method magic to get the ports forwarded for OSX and Windows people. I have a clear idea on how to do this so don't waste your time on it.

Everything will be bound to localhost for security.